+++ EnhanceQoL/Locales/enUS.lua
@@ -14,13 +14,9 @@ L["Right-Click for options"] = "Right-click for options"
 L["Lock DataPanel position (hold Shift to move)"] = "Lock DataPanel position (hold Shift to move)"
 L["Hide border"] = "Hide border"
 
--- Character/Inspect options
 L["Character & Inspect Info"] = "Character & Inspect Info"
-L["Character Frame info"] = "Character Frame info"
-L["Inspect Frame info"] = "Inspect Frame info"
-L["Item level"] = "Item level"
-L["Gem slots"] = "Gem slots"
-L["Enchants"] = "Enchants"
+L["Show on Character Frame"] = "Show on Character Frame"
+L["Show on Inspect Frame"] = "Show on Inspect Frame"
 L["Gem slot tooltip"] = "Gem slot tooltip"
 
 -- Category titles